requirements.txt
configparser==6.0.0
fastapi==0.103.0
paho-mqtt==1.6.1
psycopg2==2.9.7
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
uvicorn==0.23.2
